FastAPI_Labs/requirements.txt
scikit-learn==1.5.1
fastapi[all]==0.111.1
xgboost==2.0.3
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
tqdm==4.66.1
streamlit==1.28.1
Pillow==10.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.4
imgaug==0.4.0
tensorflow==2.15.0
matplotlib==3.8.2
